refactor(artifacts): add write
