+++ lib/artifacts/index.js
@@ -1,5 +1,6 @@
 'use strict';
 
 module.exports = {
-    ArtifactTask: require('./artifact-task')
+    ArtifactTask: require('./artifact-task'),
+    write: require('./write')
 };
