+++ lib/artifacts/write.js
@@ -10,9 +10,9 @@ const globStream = proxyquire('glob-stream', {
     })
 });
 
-const ArtifactTask = require('./artifacts').ArtifactTask;
-const TarStream = require('./streams/tar-stream');
-const CopyStream = require('./streams/copy-stream');
+const ArtifactTask = require('./artifact-task');
+const TarStream = require('../streams/tar-stream');
+const CopyStream = require('../streams/copy-stream');
 
 /**
  * Creates and writes artifact to fs.
