+++ test/write-artifact/defaults.test.js
@@ -10,7 +10,7 @@ const test = require('ava');
 const promisify = require('es6-promisify');
 const mockFs = require('mock-fs');
 
-const writeArtifact = promisify(require('../../lib/write-artifact'));
+const writeArtifact = promisify(require('../../lib/artifacts').write);
 const cwd = process.cwd();
 
 test.afterEach(() => mockFs.restore());
