+++ test/write-artifact/dest-dir.test.js
@@ -6,7 +6,7 @@ const test = require('ava');
 const promisify = require('es6-promisify');
 const mockFs = require('mock-fs');
 
-const writeArtifact = promisify(require('../../lib/write-artifact'));
+const writeArtifact = promisify(require('../../lib/artifacts').write);
 const cwd = process.cwd();
 
 test.afterEach(() => mockFs.restore());
